Modify : TeamNav페이지에서 사용되는 함수 컴포넌트 분리, 팀원배열에 map함수 속성 key 추가
