+++ README.md
@@ -8,4 +8,8 @@
 최우종
 
 TeamNav(팀 네비게이션 컴포넌트)
-현재상황 : 버튼을 눌렀을때 각각 팀원/페이지별 이동할 수 있도록 버튼 만듬. 팀 네비게이션바 상단 고정시킴
+
+최근 수정사항 :
+
+1. TeamNav.js 컴포넌트 분리함
+2. TeamMemberArray.map -> key추가
