+++ src/components/Nav/TeamMemberArray.js
@@ -1,0 +1,6 @@
+export const TeamMemberArray = [
+  { name: 'SeongSu', urlmain: '/seongsu', urllogin: '/seongsu/login' },
+  { name: 'DoYoung', urlmain: '/doyoung', urllogin: '/doyoung/login' },
+  { name: 'YuJin', urlmain: '/yujin', urllogin: '/yujin/login' },
+  { name: 'WooJong', urlmain: '/woojong', urllogin: '/woojong/login' },
+];
